Crate aws_smithy_client

A Hyper-based Smithy service client.

Feature	Description

event-stream	Provides Sender/Receiver implementations for Event Stream codegen.

rt-tokio	Run async code with the tokio runtime

test-util	Include various testing utils

native-tls	Use native-tls as the HTTP client’s TLS implementation

rustls	Use rustls as the HTTP client’s TLS implementation

client-hyper	Use hyper to handle HTTP requests

Re-exports

pub use timeout::TimeoutLayer;

Modules

bounds

This module holds convenient short-hands for the otherwise fairly extensive trait bounds required for call and friends.

conns

Type aliases for standard connection types.

erase

Type-erased variants of Client and friends.

http_connector

Default connectors based on what TLS features are active. Also contains HTTP-related abstractions that enable passing HTTP connectors around.

hyper_ext

Implementation of SmithyConnector for Hyper

never

Test connectors that never return data

retry

Retry support for aws-hyper

timeout

Timeout Configuration

Structs

Builder

A builder that provides more customization options when constructing a Client.

Client

Smithy service client.

SdkSuccess

Successful SDK Result

Enums

SdkError

Failed SDK Result