Module aws_smithy_client::bounds

This module holds convenient short-hands for the otherwise fairly extensive trait bounds required for call and friends.

The short-hands will one day be true trait aliases, but for now they are traits with blanket implementations. Also, due to compiler limitations, the bounds repeat a nubmer of associated types with bounds so that those bounds do not need to be repeated at the call site. It’s a bit of a mess to define, but should be invisible to callers.

Traits

SmithyConnector

A low-level Smithy connector that maps from http::Request to http::Response.

SmithyMiddleware

A Smithy middleware layer (i.e., factory).

SmithyMiddlewareService

A Smithy middleware service that adjusts aws_smithy_http::operation::Requests.

SmithyRetryPolicy

A Smithy retry policy.

Type Definitions

Parsed

A service that has parsed a raw Smithy response.